llm/requirements.txt
numpy 
scikit-learn
python-dotenv
datasets 
transformers
evaluate 
accelerate
torch
wandb
bitsandbytes